pyproject.toml
[project]
name = "google-meet-fast-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.3",
    "fastmcp>=2.13.1",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
